Update  remove unused login.blade.php file, and add create account link in index.blade.php

+++ app/Http/Controllers/OrganizationController.php
@@ -46,10 +46,10 @@ class OrganizationController extends Controller
             $table->id();
             $table->string('organizationName');
             $table->string('organizationAddress');
-            $table->unsignedInteger('organizationPhoneNumber');
+            $table->unsignedBigInteger('organizationPhoneNumber');
             $table->string('organizationEmail');
-            $table->unsignedInteger('organizationPanNumber')->nullable();
-            $table->unsignedInteger('organizationVatNumber')->nullable();
+            $table->unsignedBigInteger('organizationPanNumber')->nullable();
+            $table->unsignedBigInteger('organizationVatNumber')->nullable();
             $table->timestamps();
         });
